chore: add Node Releases JSON to bundle (#5443)
