+++ .gitignore
@@ -16,7 +16,6 @@ build
 public/robots.txt
 public/sitemap.xml
 public/en/feed/*.xml
-public/node-releases-data.json
 pages/en/blog/year-[0-9][0-9][0-9][0-9].md
 
 # Jest
