+++ package.json
@@ -32,7 +32,8 @@
     "test:storybook": "concurrently -P -k -s first -n \"storybook,test-storybook\" -c \"magenta,blue\" \"npm:storybook -- --ci\" \"wait-on http://localhost:6006 && test-storybook {@}\"",
     "test:storybook:snapshot": "npm run test:storybook -- -- --updateSnapshot",
     "test:storybook:watch": "npm run test:storybook -- -- --watch",
-    "test": "concurrently -s all -n \"test:unit,test:storybook\" -c \"yellow,green\" \"npm:test:unit\" \"npm:test:storybook\""
+    "test": "concurrently -s all -n \"test:unit,test:storybook\" -c \"yellow,green\" \"npm:test:unit\" \"npm:test:storybook\"",
+    "prepare": "husky install"
   },
   "dependencies": {
     "@emotion/react": "^11.11.1",
@@ -105,6 +106,7 @@
     "stylelint-selector-bem-pattern": "^2.1.1",
     "typescript": "^5.0.4",
     "user-agent-data-types": "^0.3.1",
-    "wait-on": "^7.0.1"
+    "wait-on": "^7.0.1",
+    "husky": "^8.0.0"
   }
 }
